+++ src/index.run.js
@@ -5,11 +5,12 @@
         .module('app')
         .run(runWebapp);
 
-    runWebapp.$inject = ['$log', 'loaderService'];
+    runWebapp.$inject = ['$log', 'loaderService', 'Analytics'];
 
-    function runWebapp($log, loaderService) {
+    function runWebapp($log, loaderService, Analytics) {
         $log.debug('app.runWebapp()', 'start');
         loaderService.start();
+        Analytics.pageView();
         $log.debug('app.runWebapp()', 'end');
     }